pyproject.toml
[project]
name = "stockshotgun"
version = "0.1.0"
description = "Multi-broker trading application for reverse split arbitrage"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "urwid==3.0.3",
    "python-dotenv==1.2.1",
    "pyotp==2.9.0",
    "httpx[http2]==0.28.1",
    "tastytrade==10.3.0",
    "firstrade==0.0.34",
    "schwab-py==1.4.0",
    "robin-stocks==3.4.0",
    "bbae-invest-api==0.1.3",
    "dspac-invest-api==0.1.3",
    "zendriver==0.15.0",
    "beautifulsoup4==4.14.2",
    "curl-cffi==0.13.0",
    "webull @ git+https://github.com/ImNotOssy/webull.git@587b3c7bea32a475f1596f3328cf515b4566f624",
]

[tool.setuptools.package-dir]
"" = "src"

[project.scripts]
stockshotgun = "main:main"
